Integrating java.logging

+++ src/main/java/org/openjdk/jextract/impl/MacroParserImpl.java
@@ -46,11 +46,13 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 class MacroParserImpl implements AutoCloseable {
-
+    private static final Logger LOGGER = Logger.getLogger(MacroParserImpl.class.getSimpleName());
     private final ClangReparser reparser;
     private final TreeMaker treeMaker;
     final MacroTable macroTable;
@@ -116,7 +118,8 @@ class MacroParserImpl implements AutoCloseable {
             precompiled.toFile().deleteOnExit();
             tu.save(precompiled);
             this.macro = Files.createTempFile("jextract$", cpp ? ".hpp" : ".h");
-            this.macro.toFile().deleteOnExit();
+            // if we run in debug mode we keep the file for debug purposes 
+            if (!JextractTool.DEBUG) this.macro.toFile().deleteOnExit();
             String[] patchedArgs = Stream.concat(
                 Stream.of(
                     // Avoid system search path, use bundled instead
@@ -125,10 +128,12 @@ class MacroParserImpl implements AutoCloseable {
                     // precompiled header
                     "-include-pch", precompiled.toAbsolutePath().toString()),
                 args.stream()).toArray(String[]::new);
+            LOGGER.log(Level.FINE, "Running macro parser");
             this.macroUnit = macroIndex.parse(macro.toAbsolutePath().toString(),
                     this::processDiagnostics,
                     false, //add serialization support (needed for macros)
                     patchedArgs);
+            LOGGER.log(Level.FINE, "Macro parser completed");
         }
 
         void processDiagnostics(Diagnostic diag) {
